More Useful SideBar Rewrite
