+++ src/components/Cards/MountainMapCard.tsx
@@ -16,35 +16,24 @@ import {
     Popper,
     useTheme
 } from "@mui/material";
-import { grey } from "@mui/material/colors";
 import VideocamIcon from "@mui/icons-material/Videocam";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import CloseIcon from "@mui/icons-material/Close";
-
-import { Cell, Label, Pie, PieChart } from "recharts";
-
-import callExternalAPIOnInterval from "../../hooks/callExternalAPIOnInterval";
+import ZoomInIcon from "@mui/icons-material/ZoomIn";
+import ZoomOutIcon from "@mui/icons-material/ZoomOut";
 
 const MountainMapCard = () => {
     return (
         <Box position={"relative"} width={"100%"} height={"100%"} sx={{ pointerEvents: "none" }}>
-            <Box
-                position={"absolute"}
-                top={0}
-                left={0}
-                width={"100%"}
-                height={"100%"}
-                display={"flex"}
-                flexDirection={"column"}
-                justifyContent={"space-between"}
-            >
-                <Box display={"flex"} justifyContent={"end"}>
-                    <Box pt={2} pr={2}>
-                        <LiveStreams />
-                    </Box>
-                </Box>
-                <Box pb={1} pl={1}>
-                    <LiftAndTrailStatus />
+            <Box position={"absolute"} bottom={0} left={0}>
+                <Box pb={2} pl={2} display={"flex"}>
+                    <LiveStreams />
+                    <Button variant={"contained"} sx={{ ml: 4 }}>
+                        <ZoomInIcon />
+                    </Button>
+                    <Button variant={"contained"} sx={{ ml: 4 }}>
+                        <ZoomOutIcon />
+                    </Button>
                 </Box>
             </Box>
             <SteamboatInteractiveMap />
@@ -94,176 +83,6 @@ const SteamboatInteractiveMap = () => {
             style={{ border: "none", pointerEvents: "auto" }}
             ref={mapRef}
         />
-    );
-};
-
-const LiftAndTrailStatus = () => {
-    const { VITE_TIME_INTERVAL, VITE_SKI_RESORT_ID } = import.meta.env;
-    const theme = useTheme();
-    const snowReport = callExternalAPIOnInterval(
-        VITE_TIME_INTERVAL,
-        `https://mtnpowder.com/feed?resortId=${VITE_SKI_RESORT_ID}`
-    )?.SnowReport;
-
-    const totalOpenLifts = parseInt(snowReport?.TotalOpenLifts);
-    const totalLifts = parseInt(snowReport?.TotalLifts);
-    const totalOpenTrails = parseInt(snowReport?.TotalOpenTrails);
-    const totalTrails = parseInt(snowReport?.TotalTrails);
-    const totalOpenAcres = parseInt(snowReport?.OpenTerrainAcres);
-    const totalAcres = parseInt(snowReport?.TotalTerrainAcres);
-
-    return (
-        <Box
-            p={1}
-            sx={{
-                width: "40%",
-                borderWidth: 2,
-                borderStyle: "solid",
-                backgroundColor:
-                    theme.palette.mode === "dark" ? "#121212" : theme.palette.neutral.light,
-                borderRadius: 5
-            }}
-            display={"flex"}
-            flexDirection={"row"}
-            justifyContent={"space-around"}
-        >
-            <MountainPieChart
-                chartUnit={"Trails"}
-                totalOpen={totalOpenTrails}
-                totalAmount={totalTrails}
-            />
-            <MountainPieChart
-                chartUnit={"Lifts"}
-                totalOpen={totalOpenLifts}
-                totalAmount={totalLifts}
-            />
-            <MountainPieChart
-                type={"Percent"}
-                chartUnit={"Acres"}
-                totalOpen={totalOpenAcres}
-                totalAmount={totalAcres}
-            />
-        </Box>
-    );
-};
-
-const MountainPieChart = (props: {
-    totalOpen: number;
-    totalAmount: number;
-    chartUnit: string;
-    type?: "Fraction" | "Percent";
-}) => {
-    const percentOpen = (props.totalOpen / props.totalAmount) * 100;
-    const data = [
-        { name: "open", value: props.totalOpen },
-        { name: "not open", value: props.totalAmount - props.totalOpen }
-    ];
-
-    const theme = useTheme();
-    const COLORS = [theme.palette.primary.light, grey[300]];
-
-    const labelType = props.type || "Fraction";
-    return (
-        <PieChart width={100} height={120}>
-            <Pie
-                data={data}
-                cx={"50%"}
-                cy={"50%"}
-                startAngle={90}
-                endAngle={450}
-                innerRadius={42}
-                outerRadius={50}
-                dataKey="value"
-            >
-                {data.map((_entry, index) => (
-                    <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                ))}
-                <Label
-                    width={30}
-                    position="center"
-                    content={
-                        labelType === "Fraction" ? (
-                            <FractionLabel
-                                totalOpen={props.totalOpen}
-                                totalAmount={props.totalAmount}
-                                chartUnit={props.chartUnit}
-                            />
-                        ) : (
-                            <PercentLabel percentOpen={percentOpen} chartUnit={props.chartUnit} />
-                        )
-                    }
-                ></Label>
-            </Pie>
-        </PieChart>
-    );
-};
-
-const FractionLabel = (
-    props: { totalOpen: number; totalAmount: number; chartUnit: string } & any
-) => {
-    const { cx, cy } = props.viewBox;
-    const theme = useTheme();
-    const textColor = theme.palette.mode === "dark" ? theme.palette.neutral.light : "#121212";
-    return (
-        <>
-            <text
-                x={cx - 15}
-                y={cy - 2}
-                fill={textColor}
-                textAnchor="middle"
-                fontSize={18}
-                fontWeight={700}
-            >
-                {props.totalOpen}
-            </text>
-            <text
-                x={cx + 15}
-                y={cy - 2}
-                fill={textColor}
-                textAnchor="middle"
-                fontSize={10}
-                fontWeight={500}
-            >
-                / {props.totalAmount}
-            </text>
-            <text fill={textColor} x={cx + 1} y={cy + 18} textAnchor="middle" fontSize={14}>
-                {props.chartUnit}
-            </text>
-        </>
-    );
-};
-
-const PercentLabel = (props: { percentOpen: number; chartUnit: string } & any) => {
-    const { cx, cy } = props.viewBox;
-    const theme = useTheme();
-    const textColor = theme.palette.mode === "dark" ? theme.palette.neutral.light : "#121212";
-
-    return (
-        <>
-            <text
-                x={cx - 5}
-                y={cy - 2}
-                fill={textColor}
-                textAnchor="middle"
-                fontSize={22}
-                fontWeight={700}
-            >
-                {props.percentOpen.toFixed(0)}
-            </text>
-            <text
-                x={cx + 15}
-                y={cy - 2}
-                fill={textColor}
-                textAnchor="middle"
-                fontSize={14}
-                fontWeight={500}
-            >
-                %
-            </text>
-            <text fill={textColor} x={cx + 1} y={cy + 18} textAnchor="middle" fontSize={14}>
-                {props.chartUnit}
-            </text>
-        </>
     );
 };
