+++ src/components/Dashboard.tsx
@@ -1,20 +1,13 @@
-import { Box, Divider } from "@mui/material";
+import { Box } from "@mui/material";
 
-import WeatherCard from "./Cards/WeatherCard";
-import LiftAndTrailDetailsCard from "./Cards/LiftAndTrailDetailsCard";
+import SideBar from "./Cards/SideBar";
 import MountainMapCard from "./Cards/MountainMapCard";
 
 const DashBoard = () => {
     return (
         <Box height={"100%"} display={"flex"} flexDirection={"row"}>
-            <Box width={"25%"} height={"100%"} display={"flex"} flexDirection={"column"} p={2}>
-                <Box width={"100%"} pb={2}>
-                    <WeatherCard />
-                </Box>
-                <Divider />
-                <Box flexGrow={1} width={"100%"}>
-                    <LiftAndTrailDetailsCard />
-                </Box>
+            <Box width={400} height={"100%"}>
+                <SideBar />
             </Box>
             <Box flexGrow={1} height={"100%"}>
                 <MountainMapCard />
